+++ spec/unit/lib/web/section_spec.rb
@@ -60,4 +60,11 @@ RSpec.describe Howitzer::Web::Section do
     subject { described_class.new(1, :test).capybara_context }
     it { is_expected.to eq(:test) }
   end
+
+  describe '#meta' do
+    let(:section) { described_class.new(1, :test) }
+    it { expect(section.meta).to be_an_instance_of(Howitzer::Meta::Entry) }
+    it { expect(section.meta.context).to eq(section) }
+    it { expect(section.instance_variable_get(:@meta)).to be_nil }
+  end
 end