+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [In git](https://github.com/strongqa/howitzer/compare/v2.1.1...master)
 
 ### New Features
-
+- [#241](https://github.com/strongqa/howitzer/issues/241) meta interface added. Allows tracking elements,sections, iframes on the page.
 ### Bug-fixes
 
 ## [v2.1.1](https://github.com/strongqa/howitzer/compare/v2.1.0...v2.1.1)